Refactor

+++ routes/projects/create.js
@@ -8,7 +8,7 @@ const NodeRSA = require('node-rsa');
 
 const deployRepositoryToServer = require('../../common/deployRepositoryToServer');
 
-async function ensureDeployKeyOnProject ({ db, config }, owner, repo, publicKey, authorization) {
+async function ensureDeployKeyOnProject ({ db, config }, owner, repo, authorization) {
   const deployKey = await postgres.getOne(db, `
     SELECT * FROM github_deployment_keys WHERE owner = $1 AND repo = $2
   `, [owner, repo]);
@@ -23,7 +23,7 @@ async function ensureDeployKeyOnProject ({ db, config }, owner, repo, publicKey,
       url: `${config.githubApiUrl}/repos/${owner}/${repo}/keys`,
       method: 'post',
       headers: {
-        authorization: authorization
+        authorization
       },
       data: JSON.stringify({
         key: publicKey.trim()
@@ -44,17 +44,16 @@ async function ensureDeployKeyOnProject ({ db, config }, owner, repo, publicKey,
 
   await axios({
     url: `${config.githubApiUrl}/repos/${owner}/${repo}/keys/${deployKey.github_key_id}`,
-    method: 'post',
     headers: {
-      authorization: authorization
-    },
-    data: JSON.stringify({
-      key: publicKey.trim()
-    })
+      authorization
+    }
   }).catch(error => {
-    console.log(error);
+    if (error.response.status === 404) {
+      return postgres.run(db, 'DELETE FROM github_deployment_keys WHERE id = $1', [deployKey.id])
+        .then(() => ensureDeployKeyOnProject ({ db, config }, owner, repo, authorization));
+    }
 
-    postgres.run(db, 'DELETE FROM github_deployment_keys WHERE id = $1', [deployKey.id]);
+    console.log(error);
   });
 }
 
@@ -92,14 +91,13 @@ async function createProject ({ db, config }, request, response) {
   response.write(JSON.stringify({
     ...project,
     privatekey: undefined
-  }, null, 2));
-  response.write('\n\n---\n\n');
+  }));
 
-  await ensureDeployKeyOnProject(config, body.owner, body.repo, request.headers.authorization);
+  await ensureDeployKeyOnProject({db, config}, body.owner, body.repo, request.headers.authorization);
 
   await deployRepositoryToServer({ db, config }, project, {
-    onOutput: function (data) {
-      response.write(data);
+    onOutput: function (deploymentId, data) {
+      response.write(JSON.stringify([deploymentId, data]) + '\n');
     }
   });
